backend/model/py/requirements.txt
python-dateutil
pandas_datareader
setuptools
matplotlib
yfinance
scipy
